+++ NEWS.md
@@ -6,6 +6,7 @@
 - Fixed bug where `sparsify(dt, sparsifyNAs = FALSE)` would error if dt had a factor column of all NAs (sometimes)
 - Added parameters `sparsifyCols` and `memEfficient` to `sparsify()`. `sparsifyCols` let's the user choose a subset of the inputted table's columns to be sparsified and `memEfficient = TRUE` will run a new, memory efficient version of sparsify().
 - Added parameter `alpha` to `rmsle()` and `msle()` functions to specify a custom offset instead of the deault (1).
+- Added parameter `roundbins` to `bin_data()` to allow bin values to be rounded to a specified decimal place
 
 ## mltools 0.3.5